adds unit tests for FileStreamProvider

+++ MCE2E.Controller/Providers/FileStreamProvider.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using MCE2E.Controller.Contracts;
 
 namespace MCE2E.Controller.Providers
@@ -9,13 +10,23 @@ namespace MCE2E.Controller.Providers
 
 		public Stream Get(string targetFilePath)
 		{
+			if (string.IsNullOrEmpty(targetFilePath))
+			{
+				throw new ArgumentNullException(nameof(targetFilePath));
+			}
+
 			var targetStream = new FileStream(targetFilePath, FileMode.Create, FileAccess.Write, FileShare.None);
 
 			return targetStream;
 		}
 
 		public void CleanUp(string targetFilePath)
-		{
+		{ 
+			if (string.IsNullOrEmpty(targetFilePath))
+			{
+				throw new ArgumentNullException(nameof(targetFilePath));
+			}
+
 			if (File.Exists(targetFilePath))
 				File.Delete(targetFilePath);
 		}
